feat (ip): add email list endpoint for email manager
